--- src/HB.FullStack.Database/Def/GuidEntity.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using HB.FullStack.Common.Entities;
-
-namespace HB.FullStack.Database.Def
-{
-    public abstract class GuidEntity : DatabaseEntity
-    {
-        [PrimaryKey]
-        [UniqueGuidEntityProperty(0)]
-        [CacheKey]
-        public string Guid { get; set; } = SecurityUtil.CreateUniqueToken();
-    }
-}
