--- src/HB.FullStack.Database/Def/IDatabaseEntity.cs
@@ -1,12 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using HB.FullStack.Common.Entities;
-
-namespace HB.FullStack.Database.Def
-{
-    public abstract class DatabaseEntity : Entity
-    {
-    }
-}
