--- src/HB.FullStack.Database/Def/IdDatabaseEntity.cs
@@ -1,7 +1,0 @@
-﻿namespace HB.FullStack.Database.Def
-{
-    public abstract class IdDatabaseEntity : DatabaseEntity
-    {
-        public abstract long Id { get; set; }
-    }
-}
