--- src/HB.FullStack.Database/Def/IdGenEntity.cs
@@ -1,16 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-using HB.FullStack.Common.Entities;
-using HB.FullStack.Common.IdGen;
-
-namespace HB.FullStack.Database.Def
-{
-
-    public abstract class IdGenEntity : IdDatabaseEntity
-    {
-        [PrimaryKey]
-        [EntityProperty(0)]
-        [CacheKey]
-        [LongId]
-        public override long Id { get; set; } = IDistributedIdGen.IdGen.GetId();
-    }
-}
