+++ src/HB.FullStack.Database/Entities/EntityPropertyDef.cs
